# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "whoholds backend",
    "description": "A-share top-10 shareholder network & timeline API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/person/{name}": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Person Detail",
        "operationId": "person_detail_api_person__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bucket"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/person/{name}/disambiguate": {
      "get": {
        "tags": [
          "person"
        ],
        "summary": "Disambiguate",
        "operationId": "disambiguate_api_person__name__disambiguate_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisambiguateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/company/{code}": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "Company Detail",
        "operationId": "company_detail_api_company__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/annotation": {
      "post": {
        "tags": [
          "annotation"
        ],
        "summary": "Create Annotation",
        "operationId": "create_annotation_api_annotation_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "annotation"
        ],
        "summary": "List Annotations",
        "operationId": "list_annotations_api_annotation_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnotationResponse"
                  },
                  "title": "Response List Annotations Api Annotation Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnnotationRequest": {
        "properties": {
          "op": {
            "type": "string",
            "enum": [
              "merge",
              "split",
              "bind_qid",
              "is_person"
            ],
            "title": "Op"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "user": {
            "type": "string",
            "title": "User",
            "default": "anonymous"
          }
        },
        "type": "object",
        "required": [
          "op",
          "payload"
        ],
        "title": "AnnotationRequest"
      },
      "AnnotationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "op": {
            "type": "string",
            "enum": [
              "merge",
              "split",
              "bind_qid",
              "is_person"
            ],
            "title": "Op"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "user": {
            "type": "string",
            "title": "User"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "op",
          "payload",
          "user",
          "ts"
        ],
        "title": "AnnotationResponse"
      },
      "BucketMeta": {
        "properties": {
          "bucket_idx": {
            "type": "integer",
            "title": "Bucket Idx"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "level": {
            "type": "string",
            "enum": [
              "high",
              "mid",
              "low",
              "single"
            ],
            "title": "Level"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          },
          "total_buckets": {
            "type": "integer",
            "title": "Total Buckets"
          }
        },
        "type": "object",
        "required": [
          "bucket_idx",
          "size",
          "level",
          "label",
          "evidence",
          "total_buckets"
        ],
        "title": "BucketMeta"
      },
...
      "CoholderSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "co_count": {
            "type": "integer",
            "title": "Co Count"
          },
          "is_person": {
            "type": "boolean",
            "title": "Is Person"
          }
        },
        "type": "object",
        "required": [
          "name",
          "co_count",
          "is_person"
        ],
        "title": "CoholderSummary"
      },
      "CompanyDetail": {
        "properties": {
          "stock_code": {
            "type": "string",
            "title": "Stock Code"
          },
          "stock_name": {
            "type": "string",
            "title": "Stock Name"
          },
          "available_dates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Dates"
          },
          "current_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Date"
          },
          "top10": {
            "items": {
              "$ref": "#/components/schemas/Top10Row"
            },
            "type": "array",
            "title": "Top10"
          },
          "stack_series": {
            "items": {
              "$ref": "#/components/schemas/StackSeriesPoint"
            },
            "type": "array",
            "title": "Stack Series"
          }
        },
        "type": "object",
        "required": [
          "stock_code",
          "stock_name",
          "available_dates",
          "current_date",
          "top10",
          "stack_series"
        ],
        "title": "CompanyDetail"
      },
      "CompanyHolding": {
        "properties": {
          "stock_code": {
            "type": "string",
            "title": "Stock Code"
          },
          "stock_name": {
            "type": "string",
            "title": "Stock Name"
          },
          "report_date": {
            "type": "string",
            "title": "Report Date"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "holdings": {
            "type": "integer",
            "title": "Holdings"
          },
          "pct_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct Total"
          },
          "pct_free": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct Free"
          },
          "holdings_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holdings Value"
          },
          "source_table": {
            "type": "string",
            "title": "Source Table"
          }
        },
        "type": "object",
        "required": [
          "stock_code",
          "stock_name",
          "report_date",
          "rank",
          "holdings",
          "source_table"
        ],
        "title": "CompanyHolding"
      },
...
      "DisambiguateResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "total_companies": {
            "type": "integer",
            "title": "Total Companies"
          },
          "total_buckets": {
            "type": "integer",
            "title": "Total Buckets"
          },
          "multi_company_buckets": {
            "type": "integer",
            "title": "Multi Company Buckets"
          },
          "singletons": {
            "type": "integer",
            "title": "Singletons"
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/BucketSummary"
            },
            "type": "array",
            "title": "Buckets"
          },
          "singletons_preview": {
            "items": {
              "$ref": "#/components/schemas/SingletonPreview"
            },
            "type": "array",
            "title": "Singletons Preview"
          }
        },
        "type": "object",
        "required": [
          "name",
          "total_companies",
          "total_buckets",
          "multi_company_buckets",
          "singletons",
          "buckets",
          "singletons_preview"
        ],
        "title": "DisambiguateResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "data_source": {
            "type": "string",
            "enum": [
              "entity",
              "raw_name",
              "annotated"
            ],
            "title": "Data Source"
          },
          "bucket_meta": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BucketMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "wikidata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WikidataProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyHolding"
            },
            "type": "array",
            "title": "Companies"
          },
          "total_value_series": {
            "items": {
              "$ref": "#/components/schemas/TotalValuePoint"
            },
            "type": "array",
            "title": "Total Value Series"
          },
          "coholders": {
            "items": {
              "$ref": "#/components/schemas/CoholderSummary"
            },
            "type": "array",
            "title": "Coholders"
          }
        },
        "type": "object",
        "required": [
          "name",
          "data_source",
          "companies",
          "total_value_series",
          "coholders"
        ],
        "title": "PersonDetail"
      },
...
      "SingletonPreview": {
        "properties": {
          "bucket_idx": {
            "type": "integer",
            "title": "Bucket Idx"
          },
          "stock_code": {
            "type": "string",
            "title": "Stock Code"
          },
          "stock_name": {
            "type": "string",
            "title": "Stock Name"
          }
        },
        "type": "object",
        "required": [
          "bucket_idx",
          "stock_code",
          "stock_name"
        ],
        "title": "SingletonPreview"
      },
...
      "TotalValuePoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "date",
          "value"
        ],
        "title": "TotalValuePoint"
      },
...
      "WikidataProfile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "qid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qid"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "birth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth"
          },
          "occupations": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupations"
          },
          "employer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employer"
          },
          "zh_wiki": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zh Wiki"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "WikidataProfile"
      }
    }
  }
}